Hard coding values into deleting messages from a channel

+++ index.js
@@ -1,19 +1,19 @@
 const Discord = require('discord.js');
 const client = new Discord.Client();
+const MessageHandler = require('./messageHandler.js');
 
 client.on('ready', () => {
 	console.log('Ready Sirrrrrr.');
 });
 
 client.on('message', message => {
-	console.log(message);
-	if(message.content === 'ping' ) {
-		message.reply('pong mofo');
+	if(message.channel.name === 'constitution') {
+		if(!message.author.bot) {
+			message.author.send('Yo nerd, you cannot post in ' + message.channel.name + ' as it is offlimits to the likes of you');
+			message.delete();
+		}
 	}
-	if(!message.author.bot) {
-		console.log("DELETE")
-		message.delete();
-	}
+	if(message.content.subs
 });
 
 client.login('MzQ0MzY2NDAxMjUyNzUzNDQw.DGrwUg.fp_ATYKP7Ph_C9m3531A3PYpwvs');
